fully working refactored series data manager
